+++ src/Ext/AnimType/Body.cpp
@@ -109,6 +109,7 @@ void AnimTypeExt::ExtData::LoadFromINIFile(CCINIClass* pINI)
 	this->SplashAnims.Read(exINI, pID, "SplashAnims");
 	this->SplashAnims_PickRandom.Read(exINI, pID, "SplashAnims.PickRandom");
 	this->AttachedSystem.Read(exINI, pID, "AttachedSystem", true);
+	this->AltPalette_ApplyLighting.Read(exINI, pID, "AltPalette.ApplyLighting");
 }
 
 template <typename T>
@@ -141,6 +142,7 @@ void AnimTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->SplashAnims)
 		.Process(this->SplashAnims_PickRandom)
 		.Process(this->AttachedSystem)
+		.Process(this->AltPalette_ApplyLighting)
 		;
 }